coinmarketcap 분리
