+++ bot.js
@@ -7,6 +7,7 @@ const appConfig = require('./config/config.json')
 
 const Bittrex = require('./exchange/bittrex').Bittrex
 const Korbit = require('./exchange/korbit').Korbit
+const CoinMarketCap = require('./exchange/coinmarketcap').CoinMarketCap
 
 // const coinMarketCap = require('./exchange/coinmarketcap').coinMarketCap
 // if(!process.env.TELEGRAM_BOT_TOKEN){
@@ -45,35 +46,6 @@ bittrex.run(5000)
 
 const korbit = new Korbit()
 korbit.run(5000)
-
-const CoinMarketCap = function(){
-    if(!(this instanceof CoinMarketCap)) return new CoinMarketCap()
-
-    let global_market_summary
-    let timer
-
-    this.get_coinmarketcap_global_data = function(){
-        request('https://api.coinmarketcap.com/v1/global/', function (error, response, body) {
-            if (!error && response.statusCode == 200) {
-                global_market_summary = JSON.parse(body)
-                // console.log(global_market_summary)
-            }
-        })
-    }
-    this.run = function(interval){
-        if(timer !== undefined){
-            console.log("Stop first")
-            return
-        }
-        this.get_coinmarketcap_global_data()
-        timer = setInterval(this.get_coinmarketcap_global_data, interval)
-    }
-    this.stop = function(){
-        clearInterval(timer)
-        timer = null
-    }
-    this.getSummary = () => global_market_summary
-}
 
 const coinMarketCap = new CoinMarketCap()
 coinMarketCap.run(5000)
@@ -153,7 +125,6 @@ let prev_usdt_eth = 0.0;
 let bithumb_ticker = {
 
 }
-let bittrex_markets = [];
